login register and flash messeges

+++ django/home/views.py
@@ -1,13 +1,65 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+
 
 # rooms = [
 #     {'id':1, 'name':'Lets learn python!'},
 #     {'id':2, 'name':'design with me'},
 #     {'id':3, 'name':'Frontend develper'},
 # ]
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        # print(f"Username:{username},Password:{password}")
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR password does not exist')
+    context = {'page': page}
+    return render(request, 'saleSawari/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error has occured')
+
+    return render(request, 'saleSawari/login_register.html',{'form':form})
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -26,6 +78,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'saleSawari/Room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -36,9 +89,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'saleSawari/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -48,8 +105,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'saleSawari/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here")
+
     if request.method == 'POST':
         room.delete()
         return redirect ('home')
